Create a dedicated install upgrade job
